updated bio
Motivation
I noticed a spelling error in the bio portion of the website. I checked various sources and confirmed that "interesed" is not a valid English word and the intended word here must be "interested" I know that spelling is very important when maintaining professionalism because some people view it as a indication of ones ability to be thorough. I disagree with this personally and think we are all simply humans doing our best, mistakes are inevitable, however I do believe it is worth updating the word. 

What I Did
Changed "interesed" to "interested".